Fixed bug regarding 'nonce' field overflow in the 'Block' entity - change type 'int' -> 'long'

+++ core/src/main/java/com/neemre/btcdcli4j/core/domain/Block.java
@@ -35,7 +35,7 @@ public class Block extends Entity {
 	private String merkleRoot;
 	private List<String> tx;
 	private Long time;
-	private Integer nonce;
+	private Long nonce;
 	private String bits;
 	@Setter(AccessLevel.NONE)
 	private BigDecimal difficulty;
